+++ src/app/(protected)/(lists)/mylist/_components/FavoriteList/ListControlled/DeleteList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DeleteList";
