+++ src/app/(protected)/(lists)/mylist/_components/FavoriteList/ListControlled/ModifyList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ModifyList";
